+++ packages/expo-updates/ios/EXUpdates/EXUpdatesConfig.h
@@ -2,6 +2,8 @@
 
 #import <Foundation/Foundation.h>
 
+@class EXUpdatesCodeSigningConfiguration;
+
 NS_ASSUME_NONNULL_BEGIN
 
 typedef NS_ENUM(NSInteger, EXUpdatesCheckAutomaticallyConfig) {
@@ -25,6 +27,8 @@ FOUNDATION_EXPORT NSString * const EXUpdatesConfigSDKVersionKey;
 FOUNDATION_EXPORT NSString * const EXUpdatesConfigRuntimeVersionKey;
 FOUNDATION_EXPORT NSString * const EXUpdatesConfigHasEmbeddedUpdateKey;
 FOUNDATION_EXPORT NSString * const EXUpdatesConfigExpectsSignedManifestKey;
+FOUNDATION_EXPORT NSString * const EXUpdatesConfigCodeSigningCertificateKey;
+FOUNDATION_EXPORT NSString * const EXUpdatesConfigCodeSigningMetadataKey;
 
 FOUNDATION_EXPORT NSString * const EXUpdatesConfigCheckOnLaunchValueAlways;
 FOUNDATION_EXPORT NSString * const EXUpdatesConfigCheckOnLaunchValueWifiOnly;
@@ -42,6 +46,7 @@ FOUNDATION_EXPORT NSString * const EXUpdatesConfigCheckOnLaunchValueNever;
 @property (nonatomic, readonly) NSString *releaseChannel;
 @property (nonatomic, readonly) NSNumber *launchWaitMs;
 @property (nonatomic, readonly) EXUpdatesCheckAutomaticallyConfig checkOnLaunch;
+@property (nonatomic, readonly, strong, nullable) EXUpdatesCodeSigningConfiguration *codeSigningConfiguration;
 
 @property (nullable, nonatomic, readonly) NSString *sdkVersion;
 @property (nullable, nonatomic, readonly) NSString *runtimeVersion;